packages/interception/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interception"
version = "0.1.0"
description = "Single hit-target selection for post-bounce tennis trajectories."
readme = "README.md"
authors = []
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.0",
  "curve-v3>=0.1.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
